requirements_app.txt
flask
torch
torchvision
efficientnet_pytorch
Pillow
matplotlib